update $$lte to use new AsFloat64 function

+++ internal/integration/unified/matches.go
@@ -251,19 +251,17 @@ func evaluateSpecialComparison(ctx context.Context, assertionDoc bson.Raw, actua
 
 		// Numeric values can be compared even if their types are different (e.g. if expected is an int32 and actual
 		// is an int64).
-
-		// TODO(GODRIVER-3594): If we decide to add AsDoubleOK() as a method to RawValue, this following conversion should be updated.
 		var expectedF64 float64
 		if assertionVal.Type == bson.TypeDouble {
 			expectedF64 = assertionVal.Double()
 		} else {
-			expectedF64 = float64(assertionVal.AsInt64())
+			expectedF64 = assertionVal.AsFloat64()
 		}
 		var actualF64 float64
 		if actual.Type == bson.TypeDouble {
 			actualF64 = actual.Double()
 		} else {
-			actualF64 = float64(actual.AsInt64())
+			actualF64 = actual.AsFloat64()
 		}
 
 		if actualF64 > expectedF64 {
